Limit images associated with an event
Since motion uses an event ID that is not unique across invocations, we
need to associate based on both ID and a timestamp range. For now, the
range is fixed at 10 minutes and hopefully we don't get events longer
than that.

This is still not perfect: if motion records one event then restarts and
records an event less than 10 minutes later, we'll have a problem if the
first event isn't sent to motqueser before the second one is created.
However, this shouldn't happen in practice.